+++ src/lib/rxfm/index.ts
@@ -4,3 +4,4 @@ export * from './map-to-components';
 export * from './utils';
 export * from './components';
 export * from './events';
+export * from './runtime';